Split program modules
